--- js/loaded.js
@@ -1,7 +1,0 @@
-import {SERVER_URL} from './data/data';
-
-export default class Loader {
-  static load() {
-    return fetch(`${SERVER_URL}`).then((response) => response.json());
-  }
-}
